tests: Reorganize story tests to be inside story component folder
